Add files via upload
this is Minsweeper code in C which has a lot of feature in CMD in windows:
1. has developer mode
2. coloring texts
3. type writring effect for prompting to user
4. profile covering
5. playing sounds for winner or loser (need discription)
6. high readability 
7. modular coding
8. cheat map for developer
9. etc.

if you want to play sounds, tou have to open your CMake text file and add "target_link_libraries(TARGET_NAME winmm)" to file.  be careful that I put my own sound system path. remember to change it.

made with love in IRAN